[feature] DM 리스트 > 새로운 DM 생성 > 차단 목록 화면 이동 (JND-4981)

+++ app/src/main/java/com/tosslab/jandi/app/events/RequestMoveDirectMessageEvent.java
@@ -4,9 +4,9 @@ package com.tosslab.jandi.app.events;
  * Created by justinygchoi on 2014. 9. 4..
  */
 public class RequestMoveDirectMessageEvent {
-    public int userId;
+    public long userId;
 
-    public RequestMoveDirectMessageEvent(int userId) {
+    public RequestMoveDirectMessageEvent(long userId) {
         this.userId = userId;
     }
 }
